[ADD] migration scripts for 	account_followup 	base_calendar 	project       verified nothing to do for 	knowledge 	wiki 	resource

+++ addons/email_template/migrations/6.1.1.1/pre-migration.py
@@ -41,10 +41,13 @@ def fix_email_from(cr):
         "email_template.from_account = email_template_account.id")
 
 def migrate(cr, version):
+    if not version:
+        return
     try:
         logger.info("%s called", me)
-        openupgrade.rename_columns(cr, column_renames)
-        fix_email_from(cr)
-        openupgrade.delete_model_workflow(cr, 'email_template.account')
+        if not openupgrade.column_exists(cr, 'email_template', 'user_signature'):
+            openupgrade.rename_columns(cr, column_renames)
+            fix_email_from(cr)
+            openupgrade.delete_model_workflow(cr, 'email_template.account')
     except Exception, e:
         raise osv.except_osv("OpenUpgrade", '%s: %s' % (me, e))
